Updated icons and added simple list view.
