+++ src/apps/formBuilder/components/MultipleChoice.js
@@ -33,5 +33,5 @@ const MultipleChoiceOption = ({option, active, setActive}) => (
 )
 
 const RadioToggle = ({active}) => (
-  <i className={`far fa-${active ? 'check-circle blue': 'circle-o'} radio-toggle`}/>
+  <i className={`${active ? 'fas fa-check-circle blue': 'far fa-circle'} radio-toggle`}/>
 );
