+++ src/apps/formBuilder/index.js
@@ -1,108 +1,41 @@
-import React, {Component} from 'react';
-import EditableList from './components/EditableList';
+import React from 'react';
 import formIcon from '../../images/form-icon.svg';
-import {
-    multipleChoiceForm,
-    scaledMatrixForm,
-    openResponseForm
-  } from './components/Forms';
-import {
-    MultipleChoicePreview,
-    OpenResponsePreview,
-    MatrixScalePreview
-  } from './components/Previews';
+import {SearchWithFilters as SearchInput } from '@kickup/pulse-ui/src/deprecated';
+import {Link} from 'react-router-dom';
 
 import './form-builder.less';
-import './inline-tabs.less';
 
-export default class FormBuilder extends Component {
-  constructor(props){
-    super(props);
-    this.state = {
-      name: 'Cool Form'
-    }
-    this.changeName = this.changeName.bind(this);
-  }
-  changeName(e) {
-    const name = e.target.value;
-    this.setState({name})
-  }
-  render(){
-    return (
-      <div className="wrapper" style={{paddingTop: 20}}>
-        <h2 className="header-title"><img src={formIcon} alt="Form Icon" className="header-icon"/><strong>Edit {this.state.name}</strong></h2>
-        <div className="block-flat form-builder" style={{marginBottom: 3, boxShadow: 'none', borderBottom: 'none'}}>
-          <div className="form-builder__section">
-            {/* <div className="form-builder__header">
-              <div className="circle-icon pulse-blue form-builder__header-icon">1</div>
-              <div className="form-builder__header-text">
-                <h3 className="form-builder__header-title">Form Details</h3>
-                <div className="form-builder__header-subtitle">Enter a name for your form to continue.</div>
-              </div>
-            </div> */}
-
-            <div className="form-builder__section-content">
-              <fieldset>
-                <label>Form Name</label>
-                <input className="form-control" value={this.state.name} onChange={this.changeName}/>
-              </fieldset>
-            </div>
-          </div>
-          </div>
-          <div className="inline-tabs">
-            <a className="inline-tab meta active"><i className="far fa-align-justify"/> Layout </a>
-            <a className="inline-tab meta"><i className="far fa-paint-brush"/> Styles</a>
-            <a className="inline-tab meta"><i className="far fa-cog"/> Advanced Settings</a>
-            <a className="inline-tab meta"><i className="far fa-users"/> Sharing</a>
-          </div>
-            <div className="form-builder__section">
-              <div className="row">
-                <div className="col-md-12">
-                  <div className="form-builder__editable-section block-flat">
-                    <div className="form-builder__editable-section-header">
-                      <label>Section Title</label>
-                      <input className="form-control" placeholder="Add a Section Title"/>
-                    </div>
-
-                    <div className="form-builder__editable-section-content">
-                      <EditableList
-                        items={widgets}
-                      />
-                    </div>
-                  </div>
-                  <div className="form-builder__editable-section block-flat">
-                    <div className="form-builder__editable-section-content">
-                      <EditableList
-                        items={widgets.slice(0,2)}
-                      />
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
+export default () => (
+  <div className="wrapper">
+    <h2 className="header-title"><img src={formIcon} alt="Form Icon" className="header-icon"/><strong>Forms</strong></h2>
+    <div className="block-flat">
+      <div style={{height: 40}}>
+        <SearchInput
+          placeholder="Search for a form"
+        />
       </div>
-    );
-  }
-}
-
-const widgets = [
-  {
-    name: "Multiple Choice",
-    iconClass: "multiple-choice",
-    form: multipleChoiceForm,
-    preview: MultipleChoicePreview
-  },
-  {
-    name: "Open Response",
-    iconClass: "open-responses",
-    form: openResponseForm,
-    preview: OpenResponsePreview
-  },
-  {
-    name: "Scaled Matrix",
-    iconClass: "heatmap",
-    form: scaledMatrixForm,
-    preview: MatrixScalePreview
-  }
-];
-
+      <table className="no-border" style={{marginTop: 10}}>
+        <thead className="no-border">
+          <tr>
+            <th><strong>Form</strong></th>
+            <th><strong>Actions</strong></th>
+          </tr>
+        </thead>
+        <tbody className="no-border-x no-border-y">
+          <tr>
+            <td><strong>Cool Form</strong></td>
+            <td style={{width: 150}}><Link to="/form-builder/edit" className="btn btn-primary">Edit</Link></td>
+          </tr>
+          <tr>
+            <td><strong>Other Form</strong></td>
+            <td style={{width: 150}}><Link to="/form-builder/edit" className="btn btn-primary">Edit</Link></td>
+          </tr>
+          <tr>
+            <td><strong>Nice Form</strong></td>
+            <td style={{width: 150}}><Link to="/form-builder/edit" className="btn btn-primary">Edit</Link></td>
+          </tr>
+        </tbody>
+      </table>
+    </div>
+  </div>
+)
